+++ src/lib.rs
@@ -6,6 +6,7 @@ pub mod cli;
 pub mod complete;
 pub mod config;
 pub mod cow;
+pub mod git_exclude;
 pub mod meta;
 pub mod process;
 pub mod prompt;